Fix Natspec and overall documentation

+++ contracts/buyback/YieldReserve.sol
@@ -8,23 +8,29 @@ import {IVault} from "../interfaces/IVault.sol";
 import {IOracle} from "../interfaces/IOracle.sol";
 import {Helpers} from "../libraries/Helpers.sol";
 
-/// @title YieldReserve of USDs protocol
-/// @notice The contract allows users to swap the supported stable coins against yield earned by USDs protocol
-///         It sends USDs to dripper for rebase, and to Buyback Contract for buyback.
-/// @author Sperax Foundation
+/**
+ * @title YieldReserve of USDs Protocol
+ * @notice This contract allows users to swap supported stablecoins for yield earned by the USDs protocol.
+ * It sends USDs to the dripper for rebase and to the Buyback Contract for buyback.
+ * @author Sperax Foundation
+ */
 contract YieldReserve is ReentrancyGuard, Ownable {
     using SafeERC20 for IERC20;
 
-    address public vault;
-    address public oracle;
-    address public buyback;
-    address public dripper;
-    // Percentage of USDs to be sent to Buyback 5000 means 50%
+    // Addresses of key contracts
+    address public vault; // Address of the Vault contract
+    address public oracle; // Address of the Oracle contract
+    address public buyback; // Address of the Buyback contract
+    address public dripper; // Address of the Dripper contract
+
+    // Percentage of USDs to be sent to Buyback (e.g., 5000 for 50%)
     uint256 public buybackPercentage;
 
-    mapping(address => bool) public isAllowedSrc;
-    mapping(address => bool) public isAllowedDst;
+    // Token permission mappings
+    mapping(address => bool) public isAllowedSrc; // Allowed source tokens
+    mapping(address => bool) public isAllowedDst; // Allowed destination tokens
 
+    // Events
     event Swapped(
         address indexed srcToken,
         address indexed dstToken,
@@ -50,16 +56,19 @@ contract YieldReserve is ReentrancyGuard, Ownable {
     event SrcTokenPermissionUpdated(address indexed token, bool isAllowed);
     event DstTokenPermissionUpdated(address indexed token, bool isAllowed);
 
+    // Custom error messages
     error InvalidSourceToken();
     error InvalidDestinationToken();
     error AlreadyInDesiredState();
     error TokenPriceFeedMissing();
 
-    /// @notice Constructor of the contract
-    /// @param _buyback Address of Buyback contract
-    /// @param _vault Address of Vault
-    /// @param _oracle Address of Oracle
-    /// @param _dripper Address of the dripper contract
+    /**
+     * @notice Constructor of the YieldReserve contract
+     * @param _buyback Address of the Buyback contract
+     * @param _vault Address of the Vault
+     * @param _oracle Address of the Oracle
+     * @param _dripper Address of the Dripper contract
+     */
     constructor(
         address _buyback,
         address _vault,
@@ -71,17 +80,19 @@ contract YieldReserve is ReentrancyGuard, Ownable {
         updateOracleAddress(_oracle);
         updateDripperAddress(_dripper);
 
-        /// @dev buybackPercentage is initialized to 50%
-        updateBuybackPercentage(uint256(5000));
+        // Initialize buybackPercentage to 50%
+        updateBuybackPercentage(5000);
     }
 
     // OPERATION FUNCTIONS
 
-    /// @notice Swap function to be called by front end
-    /// @param _srcToken Source / Input token
-    /// @param _dstToken Destination / Output token
-    /// @param _amountIn Input token amount
-    /// @param _minAmountOut Minimum output tokens expected
+    /**
+     * @notice Swap function to be called by frontend users
+     * @param _srcToken Source/Input token
+     * @param _dstToken Destination/Output token
+     * @param _amountIn Input token amount
+     * @param _minAmountOut Minimum output tokens expected
+     */
     function swap(
         address _srcToken,
         address _dstToken,
@@ -100,9 +111,13 @@ contract YieldReserve is ReentrancyGuard, Ownable {
 
     // ADMIN FUNCTIONS
 
-    /// @notice A function to allow or disallow a `_token`
-    /// @param _token Address of the token
-    /// @param _isAllowed If True, allow it to be used as src token / input token else don't allow
+    /**
+     * @notice Allow or disallow a specific token for use as a source/input token.
+     * @param _token Address of the token to be allowed or disallowed.
+     * @param _isAllowed If set to `true`, the token will be allowed as a source/input token; otherwise, it will be disallowed.
+     * @dev Only the contract owner can call this function.
+     * @dev If allowing a token, ensure that it has a valid price feed in the Oracle.
+     */
     function toggleSrcTokenPermission(
         address _token,
         bool _isAllowed
@@ -114,9 +129,13 @@ contract YieldReserve is ReentrancyGuard, Ownable {
         emit SrcTokenPermissionUpdated(_token, _isAllowed);
     }
 
-    /// @notice A function to allow or disallow a `_token`
-    /// @param _token Address of the token
-    /// @param _isAllowed If True, allow it to be used as src token / input token else don't allow
+    /**
+     * @notice Allow or disallow a specific token for use as a destination/output token.
+     * @param _token Address of the token to be allowed or disallowed.
+     * @param _isAllowed If set to `true`, the token will be allowed as a destination/output token; otherwise, it will be disallowed.
+     * @dev Only the contract owner can call this function.
+     * @dev If allowing a token, ensure that it has a valid price feed in the Oracle.
+     */
     function toggleDstTokenPermission(
         address _token,
         bool _isAllowed
@@ -128,10 +147,13 @@ contract YieldReserve is ReentrancyGuard, Ownable {
         emit DstTokenPermissionUpdated(_token, _isAllowed);
     }
 
-    /// @notice Emergency withdrawal function for unexpected situations
-    /// @param _token Address of the asset to be withdrawn
-    /// @param _receiver Address of the receiver of tokens
-    /// @param _amount Amount of tokens to be withdrawn
+    /**
+     * @notice Emergency withdrawal function for unexpected situations.
+     * @param _token Address of the asset to be withdrawn.
+     * @param _receiver Address of the receiver of tokens.
+     * @param _amount Amount of tokens to be withdrawn.
+     * @dev Only the contract owner can call this function.
+     */
     function withdraw(
         address _token,
         address _receiver,
@@ -142,11 +164,12 @@ contract YieldReserve is ReentrancyGuard, Ownable {
         emit Withdrawn(_token, _receiver, _amount);
     }
 
-    /// @notice Set the % of minted USDs sent Buyback
-    /// @param _toBuyback % of USDs sent to Buyback
-    /// @dev The rest of the USDs is sent to VaultCore
-    /// @dev E.g. _toBuyback == 3000 means 30% of the newly
-    ///        minted USDs would be sent to Buyback; the rest 70% to VaultCore
+    /**
+     * @notice Set the percentage of newly minted USDs to be sent to the Buyback contract.
+     * @param _toBuyback The percentage of USDs sent to Buyback (e.g., 3000 for 30%).
+     * @dev The remaining USDs are sent to VaultCore for rebase.
+     * @dev Only the contract owner can call this function.
+     */
     function updateBuybackPercentage(uint256 _toBuyback) public onlyOwner {
         Helpers._isNonZeroAmt(_toBuyback);
         Helpers._isLTEMaxPercentage(_toBuyback);
@@ -154,44 +177,58 @@ contract YieldReserve is ReentrancyGuard, Ownable {
         emit BuybackPercentageUpdated(_toBuyback);
     }
 
-    /// @notice Update Buyback contract's address
-    /// @param _newBuyBack New address of Buyback contract
+    /**
+     * @notice Update the address of the Buyback contract.
+     * @param _newBuyBack New address of the Buyback contract.
+     * @dev Only the contract owner can call this function.
+     */
     function updateBuybackAddress(address _newBuyBack) public onlyOwner {
         Helpers._isNonZeroAddr(_newBuyBack);
         buyback = _newBuyBack;
         emit BuybackAddressUpdated(_newBuyBack);
     }
 
-    /// @notice Update Oracle's address
-    /// @param _newOracle New address of Oracle
+    /**
+     * @notice Update the address of the Oracle contract.
+     * @param _newOracle New address of the Oracle contract.
+     * @dev Only the contract owner can call this function.
+     */
     function updateOracleAddress(address _newOracle) public onlyOwner {
         Helpers._isNonZeroAddr(_newOracle);
         oracle = _newOracle;
         emit OracleUpdated(_newOracle);
     }
 
-    /// @notice Update Dripper's address
-    /// @param _newDripper New address of Dripper
+    /**
+     * @notice Update the address of the Dripper contract.
+     * @param _newDripper New address of the Dripper contract.
+     * @dev Only the contract owner can call this function.
+     */
     function updateDripperAddress(address _newDripper) public onlyOwner {
         Helpers._isNonZeroAddr(_newDripper);
         dripper = _newDripper;
         emit DripperAddressUpdated(_newDripper);
     }
 
-    /// @notice Update VaultCore's address
-    /// @param _newVault New address of VaultCore
+    /**
+     * @notice Update the address of the VaultCore contract.
+     * @param _newVault New address of the VaultCore contract.
+     * @dev Only the contract owner can call this function.
+     */
     function updateVaultAddress(address _newVault) public onlyOwner {
         Helpers._isNonZeroAddr(_newVault);
         vault = _newVault;
         emit VaultAddressUpdated(_newVault);
     }
 
-    /// @notice Swap allowed src token to allowed dst token.
-    /// @param _srcToken Source / Input token
-    /// @param _dstToken Destination / Output token
-    /// @param _amountIn Input token amount
-    /// @param _minAmountOut Minimum output tokens expected
-    /// @param _receiver Receiver of the tokens
+    /**
+     * @notice Swap allowed source token for allowed destination token
+     * @param _srcToken Source/Input token
+     * @param _dstToken Destination/Output token
+     * @param _amountIn Input token amount
+     * @param _minAmountOut Minimum output tokens expected
+     * @param _receiver Receiver of the tokens
+     */
     function swap(
         address _srcToken,
         address _dstToken,
@@ -216,8 +253,7 @@ contract YieldReserve is ReentrancyGuard, Ownable {
             // Mint USDs
             IERC20(_srcToken).safeApprove(vault, _amountIn);
             IVault(vault).mint(_srcToken, _amountIn, 0, block.timestamp);
-            // No need to do slippage check as it is our contract
-            // and the vault does that.
+            // No need to do a slippage check as it is our contract, and the vault does that.
         }
         IERC20(_dstToken).safeTransfer(_receiver, amountToSend);
         _sendUSDs();
@@ -230,10 +266,13 @@ contract YieldReserve is ReentrancyGuard, Ownable {
         });
     }
 
-    /// @notice A `view` function to get estimated output
-    /// @param _srcToken Input token address
-    /// @param _dstToken Output token address
-    /// @param _amountIn Input amount of _srcToken
+    /**
+     * @notice Get an estimate of the output token amount for a given input token amount
+     * @param _srcToken Input token address
+     * @param _dstToken Output token address
+     * @param _amountIn Input amount of _srcToken
+     * @return Estimated output token amount
+     */
     function getTokenBForTokenA(
         address _srcToken,
         address _dstToken,
@@ -259,16 +298,18 @@ contract YieldReserve is ReentrancyGuard, Ownable {
 
     // UTILITY FUNCTIONS
 
-    /// @notice Sends USDs to buyback as per buybackPercentage
-    ///         and rest to VaultCore for rebase
+    /**
+     * @notice Distributes USDs to the Buyback and Dripper contracts based on buybackPercentage
+     * @dev Sends a portion of the USDs balance to the Buyback contract and the remaining to the Dripper for rebase
+     */
     function _sendUSDs() private {
         uint256 balance = IERC20(Helpers.USDS).balanceOf(address(this));
 
-        // Calculating the amount to send to Buyback based on buybackPercentage
+        // Calculate the amount to send to Buyback based on buybackPercentage
         uint256 toBuyback = (balance * buybackPercentage) /
             Helpers.MAX_PERCENTAGE;
 
-        // Remaining balance will be sent to Dripper for rebase
+        // The remaining balance is sent to the Dripper for rebase
         uint256 toDripper = balance - toBuyback;
 
         emit USDsSent(toBuyback, toDripper);
